changed the dimensions of the obstacles from double to integer

+++ src/uk/ac/soton/controller/ViewController.java
@@ -235,21 +235,21 @@ public interface ViewController {
      * @param obstacleId the ID of the obstacle.
      * @return the width of the obstacle.
      */
-    Double getPredefinedObstacleWidth(String obstacleId);
+    Integer getPredefinedObstacleWidth(String obstacleId);
 
     /**
      * Returns the height of a predefined obstacle given its ID.
      * @param obstacleId the ID of the obstacle.
      * @return the height of the obstacle.
      */
-    Double getPredefinedObstacleHeight(String obstacleId);
+    Integer getPredefinedObstacleHeight(String obstacleId);
 
     /**
      * Returns the width of a predefined obstacle given its ID.
      * @param obstacleId the ID of the obstacle.
      * @return the height of the obstacle.
      */
-    Double getPredefinedObstacleLength(String obstacleId);
+    Integer getPredefinedObstacleLength(String obstacleId);
 
     /**
      * Adds an obstacle to the list of predefined obstacles
@@ -258,7 +258,7 @@ public interface ViewController {
      * @param width the width of the new obstacle.
      * @param height the height of the new obstacle.
      */
-    void addObstacleToList(String id, Double length, Double width, Double height);
+    void addObstacleToList(String id, Integer length, Integer width, Integer height);
 
     /**
      * Deletes a specified object from the list of predefined obstacles.
